Modularize user login/logout routes in separate file
